Integration tests for BoardController are finished for all endpoints

+++ TrelloProject/TrelloProject.BLL/Services/BoardDTOService.cs
@@ -29,8 +29,16 @@ namespace TrelloProject.BLL.Services
 
         public void DeleteBoardDTO(int id)
         {
-            _boardRepository.GetBoard(id);
-            _boardRepository.Delete(id);
+            BoardDTO boardDTO = _boardRepository.GetBoard(id);
+            if(boardDTO == null)
+            {
+                throw new NullReferenceException("The item with ID=" + id + " does not exist");
+            }
+            else
+            {
+                _boardRepository.Delete(id);
+            }
+            
         }
 
         public List<BoardDTO> GetAllBoardsDTO()
